+++ modules/types/src/cpp/ArrayOf_PromoteType.cpp
@@ -276,8 +276,7 @@ ArrayOf::promoteType(Class dstClass, stringVector fNames)
     // Do nothing for promoting to same class (no-op).
     if (isSparse()) {
         dp = dp->putData(dstClass, dp->dimensions,
-            TypeConvertSparseDynamicFunction(
-                dp->dataClass, dp->dimensions[0], dp->dimensions[1], dp->getData(), dstClass),
+            TypeConvertSparseDynamicFunction(dstClass, dp->dimensions[0], dp->dimensions[1], dp->getData(), dp->dataClass),
             true);
         return;
     }