+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.4.7 (2020-07-XX)
+# 0.4.7 (2020-07-31)
 
   * [#311](http://github.com/Nelson-numerical-software/nelson/issues/311): betainc builtin: Incomplete beta function.